Fix parsing and displaying Graphite function descriptions
